adding the route POST '/v1/backups/deposits/list' with paginations and filters

+++ internal/store_vault_service/get_balances_service.go
@@ -36,7 +36,7 @@ func GetBalances(
 		resDeposits[i] = &backupBalance.BackupDeposit{
 			Recipient:        deposit.Recipient,
 			EncryptedDeposit: deposit.EncryptedDeposit,
-			BlockNumber:      deposit.BlockNumber,
+			BlockNumber:      uint64(deposit.BlockNumber),
 			CreatedAt:        deposit.CreatedAt,
 		}
 	}
